Fixup completion in CIDER
few issues:
- CIDER uses dumb completion in the repl for some reason. Need to use
'company-indent-or-complete-common' as the 'cider-repl-tab-command'
- set TAB to be ("TAB" . company-indent-or-complete-common)
globally.

Was having interesting issue: autocomplete at the repl sucked, but if
invoking company-complete-common manually it worked fine. However, was
getting no completion in source buffers just a "no completions found"
error. Still not exactly sure why this solved that but cargo cults
abound. Possible it had to do with the capf cache and it gets really
hairy in there at that point.

Also added a package to show the docstrings in company autocomplete.